% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BarnacleGooseData}
\alias{BarnacleGooseData}
\title{An example dataset from the Vigilance Behaviour in Barnacle Geese (1997) video produced for the Association for the Study of Animal Behaviour  .}
\format{
A data frame with 12 rows and 5 variables:
\describe{
\item{Flock}{Index indicating which video the data were taken from}
\item{FlockSize}{The number of animals in the flock}
\item{TotalHeadsUp}{How many times the animal looked up}
\item{TimeSecHeadUp}{The total duration (s) the animal was vigilant}
\item{Partner}{Which partner collected the data}
...
}
}
\source{
https://www.asab.org/
}
\usage{
BarnacleGooseData
}
\description{
This dataset is based off data collected from this video: https://www.youtube.com/watch?v=oeJHW5n-PXY
}
\keyword{datasets}
