% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiplotAddSpectrograms.R
\name{BiplotAddSpectrograms}
\alias{BiplotAddSpectrograms}
\title{Function to create a principal components biplot and add spectrograms}
\usage{
BiplotAddSpectrograms(input.dir.Focal = "FocalRecordings")
}
\description{
Function to create a principal components biplot and add spectrograms
}
