% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CameraTrapAnnotations.R
\name{CameraTrapAnnotations}
\alias{CameraTrapAnnotations}
\title{Function to create a table of available camera trap annotations by season}
\usage{
CameraTrapAnnotations(season = 4)
}
\arguments{
\item{season}{which season to download; options are 1-4, 6-8,10,11}
}
\value{

}
\description{
Function to create a table of available camera trap annotations by season
}
