% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CameraTrapDataAccess.R
\name{CameraTrapDataAccess}
\alias{CameraTrapDataAccess}
\title{Function to access Snapshot Serengeti camera trap photos on the Labeled Information Library of Alexandria: Biology and Conservation http://lila.science/datasets/snapshot-serengeti}
\usage{
CameraTrapDataAccess(
  urlpath = "https://lilablobssc.blob.core.windows.net/snapshotserengeti-unzipped/",
  season = 4,
  AnimalID = "leopard",
  NumPhotos = 5,
  create.dir = TRUE
)
}
\arguments{
\item{create.dir}{}
}
\value{

}
\description{
Function to access Snapshot Serengeti camera trap photos on the Labeled Information Library of Alexandria: Biology and Conservation http://lila.science/datasets/snapshot-serengeti
}
