% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CameraTrapDataCollection.R
\name{CameraTrapDataCollection}
\alias{CameraTrapDataCollection}
\title{Function to print each photograph and collect data on the time the photo was taken.}
\usage{
CameraTrapDataCollection(
  inputfile = CombinedAnimalDF,
  rowstart = 1,
  dataframe.cont = FALSE
)
}
\arguments{
\item{inputfile}{the dataframe output by the CameraTrapDataAccess function}

\item{rowstart}{if data collection was interrupted using the command 'break' which row to use to resume}

\item{dataframe.cont}{logical; change to 'TRUE' if data collection was interrupted}
}
\value{

}
\description{
Function to print each photograph and collect data on the time the photo was taken.
}
