% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CameraTrapDataOutput}
\alias{CameraTrapDataOutput}
\title{A dataframe that is the output of the 'CameraTrapDataAccess' function}
\format{
A data frame with 10 rows and 5 variables:
\describe{
\item{category_id}{A factor indicating which species was detected}
\item{season}{A factor indicating the season the image was taken}
\item{location}{The camera trap location}
\item{filename}{The file path of the downloaded image}
...
}
}
\source{
Swanson AB, Kosmala M, Lintott CJ, Simpson RJ, Smith A, Packer C (2015) Snapshot Serengeti, high-frequency annotated camera trap images of 40 mammalian species in an African savanna. Scientific Data 2: 150026.

http://lila.science/datasets/snapshot-serengeti
}
\usage{
CameraTrapDataOutput
}
\description{
This is a sample dataframe when the 'CameraTrapDataAccess' function is run using defaults
}
\keyword{datasets}
