% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CombinedDFTimes}
\alias{CombinedDFTimes}
\title{A simulated dataset from two censuses conducted along a 500 m transect}
\format{
A data frame with 78 rows and 6 variables:
\describe{
\item{Species}{Which species was detected}
\item{DistanceAlongTransect}{How far along the transect line the animal was detected}
\item{PerpendicularDistance}{Perpendicular distance from the transect line}
\item{MethodDetection}{How the animal was detected}
\item{Site}{A variable indicating which transect line}
\item{Partner}{Which partner collected the data}
...
}
}
\source{
NA
}
\usage{
CombinedDFTimes
}
\description{
In this dataset students indicate all species that they detect along a 500-m transect.
}
\keyword{datasets}
