% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CombinedPartnerDFTimes}
\alias{CombinedPartnerDFTimes}
\title{A dataframe containing camera trap data annotations from Snapshot serengeti of leopards and cheetahs}
\format{
A data frame with 265 rows and 5 variables:
\describe{
\item{category_id}{A factor indicating which species was detected}
\item{season}{A factor indicating the season the image was taken}
\item{location}{The camera trap location}
\item{filename}{The file path of the downloaded image}
\item{Time }{The time (in hours on the 24-hour clock) that the image was taken}
...
}
}
\source{
Swanson AB, Kosmala M, Lintott CJ, Simpson RJ, Smith A, Packer C (2015) Snapshot Serengeti, high-frequency annotated camera trap images of 40 mammalian species in an African savanna. Scientific Data 2: 150026.

http://lila.science/datasets/snapshot-serengeti
}
\usage{
CombinedPartnerDFTimes
}
\description{
This is an example of how the dataset will look after students collect data using both 'CameraTrapDataAccess' and 'CameraTrapDataCollection' functions
}
\keyword{datasets}
