% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrelatedDataSimulationFunction.R
\name{CorrelatedDataSimulationFunction}
\alias{CorrelatedDataSimulationFunction}
\title{Title}
\usage{
CorrelatedDataSimulationFunction(
  N = 100,
  CorrelationCoefficient = 0.83,
  MaleMeanBodyWeight = 125,
  MaleReproductiveSuccess = 3
)
}
\arguments{
\item{MaleReproductiveSuccess}{}
}
\value{

}
\description{
Title
}
