% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DeerSpeciesAcousticFeatures}
\alias{DeerSpeciesAcousticFeatures}
\title{A toy dataset of red deer, sika deer and red x sika hybrid vocalizations.}
\format{
A data frame with 12 rows and 4 variables:
\describe{
\item{Class}{RedDeer, Hybrids Sikas}
\item{Duration}{Simulated duration of vocalizations}
\item{MinFrequency}{Simulated minimum frequency of vocalizations}
\item{MaxFrequency}{Simulated maximum frequency of vocalizations}
...
}
}
\source{
Long, A. M., N. P. Moore, and T. J. Hayden. “Vocalizations in red deer (Cervus elaphus), sika deer (Cervus nippon), and red× sika hybrids.” Journal of Zoology 244.1 (1998): 123-134.
}
\usage{
DeerSpeciesAcousticFeatures
}
\description{
This dataset contains four vocalizations per individual and three aspects of the vocalizations: the duration, the minimum frequency and the maximum frequency of the vocalization
}
\keyword{datasets}
