% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MaleDeerRoarDF}
\alias{MaleDeerRoarDF}
\title{A simulated data set indicating male red deer roars per minute as a function of mating season or harem-holding status.}
\format{
A data frame with 20 rows and 2 variables:
\describe{
\item{MaleCategory}{MaleCategory, PreMatingSeason, HaremHolders, NotHaremHolders or PostMatingSeason"}
\item{RoarsPerMinute}{Mean number of roars per minute}
...
}
}
\source{
Clutton-Brock, T. H., & Albon, S. D. (1979). The roaring of red deer and the evolution of honest advertisement. Behaviour, 69(3-4), 145-170.
}
\usage{
MaleDeerRoarDF
}
\description{
This dataset contains two variables: MaleCategory and RoarsPerMinute
}
\keyword{datasets}
