% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MeerkatFocalData}
\alias{MeerkatFocalData}
\title{A dataframe containing simulated focal data from meerkat observations}
\format{
A data frame with 28 rows and 6 variables:
\describe{
\item{StartTimeMin}{The start time (minute only) that the behavior started}
\item{StartTimeSec}{The start time (seconds only) that the behavior started}
\item{BehaviorCode}{Which behavior the animal engaged in}
\item{TimeSeconds}{The cumulative time in seconds that the behavior was observed}
\item{SecondsEngagedinBehavior }{The total time (in seconds) the animal was observed engaging in the behavior}
\item{Partner}{Which partner collected the data}
...
}
}
\source{
Hammond 2019, Vigilance behaviour in meerkats, ASAB Education


}
\usage{
MeerkatFocalData
}
\description{
This simulated data is based on exercises from 'Vigilance behavour in meerkats'
}
\keyword{datasets}
