% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MeerkatScanData}
\alias{MeerkatScanData}
\title{A dataframe containing simulated scan data from meerkat observations}
\format{
A data frame with 28 rows and 6 variables:
\describe{
\item{Time}{The start time (in seconds) of the scan}
\item{Vigilant}{How many animals were vigilant during the scan}
\item{NotVigilant}{How many animals weren't viginlant during the scan}
\item{OutOfSight}{Number of animals out of sight}
\item{Treatment }{A factor indicating whether there was no predator, an aerial predator or a terrestrial predator}
\item{Partner}{Which partner collected the data}
...
}
}
\source{
Hammond 2019, Vigilance behaviour in meerkats, ASAB Education


}
\usage{
MeerkatScanData
}
\description{
This simulated data is based on exercises from 'Vigilance behavour in meerkats'
}
\keyword{datasets}
