% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFCCFunction.R
\name{MFCCFunction}
\alias{MFCCFunction}
\title{Title}
\usage{
MFCCFunction(
  input.dir,
  min.freq = 200,
  max.freq = 2000,
  n.windows = 9,
  num.cep = 12
)
}
\arguments{
\item{num.cep}{}
}
\value{

}
\description{
Title
}
