% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFCCFunctionSite.R
\name{MFCCFunctionSite}
\alias{MFCCFunctionSite}
\title{Title}
\usage{
MFCCFunctionSite(
  input.dir,
  min.freq = 0,
  max.freq = 20000,
  n.windows = 9,
  num.cep = 12
)
}
\arguments{
\item{num.cep}{}
}
\value{

}
\description{
Title
}
