% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrogramFunction.R
\name{SpectrogramFunction}
\alias{SpectrogramFunction}
\title{A function to calculate spectrograms over a set of focal recordings}
\usage{
SpectrogramFunction(
  input.dir,
  min.freq = 500,
  max.freq = 2500,
  Colors = "BW",
  downsample.new = 16000
)
}
\arguments{
\item{input.dir}{The location of the .wav files}

\item{min.freq}{The minimum frequency (Hz) of the signal of interest}

\item{max.freq}{The maximum frequency (Hz) of the signal of interest}

\item{Colors}{Whether to return a black and white or color spectrogram}

\item{downsample.new}{<- If downsampling is desired, which frequency (Hz) to use for downsampling}
}
\value{

}
\description{
A function to calculate spectrograms over a set of focal recordings
}
\examples{
\donttest{SpectrogramFunction(input.dir = "FocalRecordings",
min.freq = 500,max.freq=2500)}
}
