% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrogramFunctionSite.R
\name{SpectrogramFunctionSite}
\alias{SpectrogramFunctionSite}
\title{Title}
\usage{
SpectrogramFunctionSite(input.dir, min.freq = 500, max.freq = 2500)
}
\arguments{
\item{max.freq}{}
}
\value{

}
\description{
Title
}
