% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrogramSingle.R
\name{SpectrogramSingle}
\alias{SpectrogramSingle}
\title{Title}
\usage{
SpectrogramSingle(
  sound.file,
  min.freq = 500,
  max.freq = 8000,
  Colors = "BW",
  downsample = TRUE
)
}
\arguments{
\item{downsample}{}
}
\value{

}
\description{
Title
}
