% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiplotAddSpectrograms.R
\name{BiplotAddSpectrograms}
\alias{BiplotAddSpectrograms}
\title{Title}
\usage{
BiplotAddSpectrograms(
  input.dir.Focal = "FocalRecordings",
  xmax = 15,
  xmin = -15,
  ymax = 10,
  ymin = -10
)
}
\arguments{
\item{ymin}{}
}
\value{

}
\description{
Title
}
