% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CameraTrapDataAccess.R
\name{CameraTrapDataAccess}
\alias{CameraTrapDataAccess}
\title{Snapshot Serengeti camera trap access}
\usage{
CameraTrapDataAccess(
  urlpath = "https://lilablobssc.blob.core.windows.net/snapshotserengeti-unzipped/",
  season = list(1, 2),
  AnimalID = "leopard",
  NumPhotos = 5,
  create.dir = ""
)
}
\arguments{
\item{urlpath}{location of Snapshot Serengeti database}

\item{season}{which season to query; options are 1-4, 6-8,10,11}

\item{AnimalID}{which animal photos to download}

\item{NumPhotos}{the number of photos to download}

\item{create.dir}{logical; create a new directory (TRUE) or save to current temp directory (FALSE); to see temp directory use tempdir()}
}
\value{
a dataframe with image metadata and downloaded images
}
\description{
Function to access Snapshot Serengeti camera trap photos on the Labeled Information Library of Alexandria: Biology and Conservation http://lila.science/datasets/snapshot-serengeti
}
\examples{
\donttest{CombinedAnimalDF <- CameraTrapDataAccess(urlpath=
'https://lilablobssc.blob.core.windows.net/snapshotserengeti-unzipped/',
season= list(1,2,3),AnimalID='leopard', NumPhotos= 50,create.dir=TRUE)}
}
