% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CameraTrapDataAccess.R
\name{CameraTrapDataAccess}
\alias{CameraTrapDataAccess}
\title{Function to access Snapshot Serengeti camera trap photos on the Labeled Information Library of Alexandria: Biology and Conservation http://lila.science/datasets/snapshot-serengeti}
\usage{
CameraTrapDataAccess(
  urlpath = "https://lilablobssc.blob.core.windows.net/snapshotserengeti-unzipped/",
  season = list(1, 2),
  AnimalID = "leopard",
  NumPhotos = 5,
  create.dir = TRUE
)
}
\arguments{
\item{urlpath}{location of Snapshot Serengeti database}

\item{season}{which season to query; options are 1-4, 6-8,10,11}

\item{AnimalID}{which animal photos to download}

\item{NumPhotos}{the number of photos to download}

\item{create.dir}{logical; create a new directory (TRUE) or save to current working directory (FALSE)}
}
\value{

}
\description{
Function to access Snapshot Serengeti camera trap photos on the Labeled Information Library of Alexandria: Biology and Conservation http://lila.science/datasets/snapshot-serengeti
}
