% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrogramFunction.R
\name{SpectrogramFunction}
\alias{SpectrogramFunction}
\title{Title}
\usage{
SpectrogramFunction(input.dir, min.freq = 500, max.freq = 2500, Colors = "BW")
}
\arguments{
\item{Colors}{}
}
\value{

}
\description{
Title
}
