% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectrogramFunctionSite.R
\name{SpectrogramFunctionSite}
\alias{SpectrogramFunctionSite}
\title{Function to create spectrograms for soundscape recordings}
\usage{
SpectrogramFunctionSite(
  input.dir,
  min.freq = 500,
  max.freq = 2500,
  Colors = "BW"
)
}
\arguments{
\item{input.dir}{The location of the .wav files}

\item{min.freq}{The minimum frequency (Hz) of the signal of interest}

\item{max.freq}{The maximum frequency (Hz) of the signal of interest}

\item{Colors}{Whether to return a black and white or color spectrogram}
}
\value{

}
\description{
Function to create spectrograms for soundscape recordings
}
\examples{
\donttest{SpectrogramFunctionSite(input.dir = "SoundscapeRecordings",
min.freq = 0,max.freq=20000)}
}
